prepare autoupgrade script, wip server url, update and reduce mount points

+++ scripts/start.sh
@@ -9,13 +9,20 @@ if [ ! -f "$fileLocalSettings" ]; then
   WIKI_ADMIN_PASS=$(openssl rand -base64 32)
   echo "$WIKI_ADMIN_PASS" > /data/wikisysop_password.txt
   cd "$basePath"; php maintenance/install.php --dbserver "$DB_HOST" --dbport "$DB_PORT" --dbname "$DB_NAME" --dbuser "$DB_USER" --dbpass "$DB_PASSWORD" --pass "$WIKI_ADMIN_PASS" --scriptpath /bluespice "$WIKI_NAME" "$WIKI_ADMIN"
-  mv ${basePath}/LocalSettings.php /data/LocalSettings.php
-  ln -s /data/LocalSettings.php ${basePath}/LocalSettings.php
 
-  echo "require \"\$IP/LocalSettings.BlueSpice.php\";" >> ${basePath}/LocalSettings.php
-  echo "wfLoadExtension('BlueSpiceExtensions/ExtendedSearch');" >> ${basePath}/LocalSettings.php
-  echo "wfLoadExtension('BlueSpiceExtensions/UniversalExport');" >> ${basePath}/LocalSettings.php
-  echo "wfLoadExtension('BlueSpiceExtensions/UEModulePDF');" >> ${basePath}/LocalSettings.php
+  #post install
+  if [ -f "$fileLocalSettings" ]; then
+    mv ${basePath}/LocalSettings.php /data/LocalSettings.php
+    ln -s /data/LocalSettings.php ${basePath}/LocalSettings.php
+
+    echo "require \"\$IP/LocalSettings.BlueSpice.php\";" >> ${basePath}/LocalSettings.php
+    echo "wfLoadExtension('BlueSpiceExtensions/ExtendedSearch');" >> ${basePath}/LocalSettings.php
+    echo "wfLoadExtension('BlueSpiceExtensions/UniversalExport');" >> ${basePath}/LocalSettings.php
+    echo "wfLoadExtension('BlueSpiceExtensions/UEModulePDF');" >> ${basePath}/LocalSettings.php
+  else
+    echo "Error occured: installation not successfull, LocalSettings.php is missing"
+  fi
+
 else
   ln -s /data/LocalSettings.php ${basePath}/LocalSettings.php
 fi
